GEM: Squash of 11 commits (Making Python 2.0 & Python 3.0 exceptions similar):
    Start adding work to make Python 2.0 exception handling the same as Python 3.0 exception handling
    Work on making Python 2.0 & Python 3.0 exceptions similar
    Cleanup of exceptions
    Make exception handling Python 2.0 same as Python 3.0
    Improve exception handling to distinguish between reraise & raising
    More Exception cleanup
    More exception handling to make Python 2.0 work like Python 3.0 (also found a bug in with stack traces for with handling in Python 3.0)
    Python 2.0 exceptions now producing same tracebacks as Python 3.0 (also now have ExceptionChain test)
    More exception cleanup
    rename exception functions to *_clause (and make the built_in)
    More exceptions fixes (can't easily get 'finally' clause in Python 2.* to work easily)

+++ work/run.sh
@@ -36,6 +36,9 @@ Main_py=../Tremolite/Main.py
 Main_py=../Quartz/Main.py
 Main_py=../Sapphire/Main.py
 Main_py=../Topaz/Main.py
+#Main_py=test3.py
+
+show=2
 
 command="python $Main_py"
 commandO="python -O $Main_py"
@@ -52,7 +55,7 @@ y
 END
 
 echo -en '\E[H\E[J'
-cat 2
+cat $show
 
 while :
 do
@@ -62,15 +65,28 @@ do
         :
     else
         mv $tmp2 2
-        echo -en '\E[H\E[J'
-        tail -60 2
+
+        if [ $show = 2 ]; then
+            echo -en '\E[H\E[J'
+            tail -60 2
+        fi
     fi
 
     $commandO <$tmp1 >&$tmp3
     mv $tmp3 2o
 
     $command3 <$tmp1 >&$tmp3
-    mv $tmp3 3
+    if cmp -s $tmp3 3
+    then
+        :
+    else
+        mv $tmp3 3
+
+        if [ $show = 3 ]; then
+            echo -en '\E[H\E[J'
+            tail -60 3
+        fi
+    fi
 
     $command3O <$tmp1 >&$tmp3
     mv $tmp3 3o
